added notifications to the postgres dastabase for orders and franchises

+++ priv/repo/migrations/20200725140903_create_product.exs
@@ -4,7 +4,7 @@ defmodule Haberdash.Repo.Migrations.CreateProduct do
   def change do
     create table(:product, primary_key: false) do
       add :name, :string
-      add :price, :decimal
+      add :price, :integer
       add :description, :text
       add :franchise_id, references(:franchise, on_delete: :delete_all, type: :binary_id)
       add :id, :binary_id, primary_key: true
