+++ priv/repo/migrations/20200818212019_create_accessories.exs
@@ -5,7 +5,7 @@ defmodule Haberdash.Repo.Migrations.CreateAccessories do
     create table(:accessories, primary_key: false) do
       add :id, :binary_id, primary_key: true
       add :name, :string
-      add :price, :decimal, default: 0
+      add :price, :integer, default: 0
       add :max_quantity, :integer, default: 2
       add :franchise_id, references(:franchise, type: :uuid, on_delete: :nothing)
       add :description, :string
